Aquarium
Fishes in an aquarium require a certain water pH level, so aquarium hobbyists must regularly measure the water’s pH. The value range must be suitable for fishes and cannot vary too much. In this task, we shall create a program that defines the water’s suitability for zebra fishes.