Noughts-and-crosses, barebones version
Let's implement a noughts and crosses game. We begin by creating a version that only allows adding marks (noughts or crosses) on the game board, but does not check when the game ends. The implementation of the program thus always ends after nine turns, independently of whether either player manages to fill a row with their mark earlier.