The longest substring in order
Implement the function longest_substring_in_order, which takes a string as its parameteri and searches for the longest substring with its characters in alphabetic order and then returns it. You can assume the string contains only lower-case letters, so you can compare the alphabetical order of the strings by using the comparison operator <, for instance.

If the string contains several equally long substrings in alphabetic order, the program returns the substring that is the closest to the beginning of the string.