OS library
After years of collection, you have managed to put together a large music collection including thousands of music files, but sadly, at the starting phases of your collection, you had decided to use a different format for naming the files than later on.

Approximately 13000 of the music files were named like you wish all the files to be named, meaning:

song_name-artist.mp3
However, the folder still has approximately 2000 files named differently:

number-artist-song_name.mp3
Create the function fix_filenames, which has the name of the folder where the changed files are located as a parameter. The function changes the names of every file in the wrong format to the correct format and leaves the names of all the other files as they were. The function does not print anything on the screen.